Add files via upload
a modern, responsive educational website designed to showcase courses, tutorials, and tech resources. The site is built with HTML, CSS, and JavaScript, featuring clean navigation, interactive elements, and a visually appealing layout. It aims to provide a smooth user experience for students and tech enthusiasts, making it easy to explore learning materials and engage with the platform. This project highlights best practices in front-end development, responsive design, and user-friendly web interfaces.